post

+++ app/Http/Controllers/TasksController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 
-use App\Task;
+use App\Http\Controllers\Controller;
 
 class TasksController extends Controller
 {
@@ -13,19 +13,30 @@ class TasksController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+ public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);}
-
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            $data += $this->counts($user);
+            return view('users.show', $data);
+        }else {
+            return view('welcome');
+        }
+    }
+    
     /**
      * Show the form for creating a new resource.
      *
      * @return \Illuminate\Http\Response
      */
+
     public function create()
     {
         $task = new Task;
@@ -45,13 +56,14 @@ class TasksController extends Controller
             'status' => 'required|max:10', 
             'content' => 'required|max:191',
         ]);
-        $task = new Task;
-        $task->status = $request->status; 
-        $task->content = $request->content;
-        $task->save();
+        $request->user()->tasks()->create([
+            'status'=> $request->status,
+            'content' => $request->content,
+            
+        ]);
 
-        return redirect('/');
-     }
+        return redirect()->back();
+    }
 
     /**
      * Display the specified resource.
@@ -113,10 +125,12 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+        $task = \App\Task::find($id);
 
-        return redirect('/');
-        //
+        if (\Auth::user()->id === $task->user_id) {
+            $task->delete();
+        }
+
+        return redirect()->back();
     }
 }
